norminette

+++ printf/ft_printf.h
@@ -17,11 +17,11 @@
 # include <stdarg.h>
 
 int	ft_d(int n);
-int	ft_p(unsigned long long ptr);
+int	ft_p(unsigned long ptr);
 void	ft_putchar_fd(char c, int fd);
 int	ft_s(char *str);
 int	ft_u(unsigned int n);
-int	ft_x(unsigned int num, const char format);
+int	ft_x(unsigned long num, const char format); 
 int	ft_c(int c);
 int	ft_printf(const char *format, ...);
